I turned the site into a PHP site.
The PHP site is going to be much easier to manage once the website
starts growing. I also updated the jQuery library to the latest version,
moved all the JavaScript to the script.js file, and removed the
JavaScript code that was not working. Finally, the site's content and
navigation bar were updated too.